Text.URI

Description

This is a modern library for working with URIs as per RFC 3986:

https://tools.ietf.org/html/rfc3986

This module is intended to be imported qualified, e.g.:

import Text.URI (URI)
import qualified Text.URI as URI

See also Text.URI.Lens for lens, prisms, and traversals; see Text.URI.QQ for quasi-quoters for compile-time validation of URIs and refined text components.

Data types

data URI

Uniform resource identifier (URI) reference. We use refined Text (RText l) here because information is presented in human-readable form, i.e. percent-decoded, and thus it may contain Unicode characters.

Constructors

URI

Fields

uriScheme :: Maybe (RText 'Scheme)

URI scheme, if Nothing, then the URI reference is relative

uriAuthority :: Either Bool Authority

Authority component in Right or a Bool value in Left indicating if uriPath path is absolute (True) or relative (False); if we have an Authority component, then the path is necessarily absolute, see isPathAbsolute

Note: before version 0.1.0.0 type of uriAuthority was Maybe Authority

uriPath :: Maybe (Bool, NonEmpty (RText 'PathPiece))

Nothing represents the empty path, while Just contains an indication Bool whether the path component has a trailing slash, and the collection of path pieces NonEmpty (RText PathPiece).

Note: before version 0.2.0.0 type of uriPath was [RText PathPiece].

uriQuery :: [QueryParam]

Query parameters, RFC 3986 does not define the inner organization of query string, so we deconstruct it following RFC 1866 here

uriFragment :: Maybe (RText 'Fragment)

Fragment, without #

mkURI :: MonadThrow m => Text -> m URI

Construct a URI from Text. The input you pass to mkURI must be a valid URI as per RFC 3986, that is, its components should be percent-encoded where necessary. In case of parse failure ParseException is thrown.

This function uses the parser parser under the hood, which you can also use directly in a Megaparsec parser.

mkURIBs :: MonadThrow m => ByteString -> m URI

Construct a URI from ByteString. The input you pass to mkURIBs must be a valid URI as per RFC 3986, that is, its components should be percent-encoded where necessary. In case of parse failure ParseExceptionBs is thrown.

This function uses the parserBs parser under the hood, which you can also use directly in a Megaparsec parser.

Since: 0.3.3.0

emptyURI :: URI

The empty URI.

Since: 0.2.1.0

makeAbsolute :: RText 'Scheme -> URI -> URI

Make a given URI reference absolute using the supplied RText Scheme if necessary.

isPathAbsolute :: URI -> Bool

Return True if path in a given URI is absolute.

Since: 0.1.0.0

relativeTo

Arguments

:: URI
Reference URI to make absolute

-> URI
Base URI

-> Maybe URI
The target URI

relativeTo reference base makes the reference URI absolute resolving it against the base URI.

If the base URI is not absolute itself (that is, it has no scheme), this function returns Nothing.

See also: https://tools.ietf.org/html/rfc3986#section-5.2.

Since: 0.2.0.0

data Authority

Authority component of URI.

Constructors

Authority

Fields

authUserInfo :: Maybe UserInfo

User information

authHost :: RText 'Host

Host

authPort :: Maybe Word

Port number

data UserInfo

User info as a combination of username and password.

Constructors

UserInfo

Fields

uiUsername :: RText 'Username

Username

uiPassword :: Maybe (RText 'Password)

Password, Nothing means that there was no : character in the user info string

data QueryParam

Query parameter either in the form of flag or as a pair of key and value. A key cannot be empty, while a value can.

Constructors

QueryFlag (RText 'QueryKey)
Flag parameter

QueryParam (RText 'QueryKey) (RText 'QueryValue)
Key–value pair

newtype ParseException

Parse exception thrown by mkURI when a given Text value cannot be parsed as a URI.

Constructors

ParseException (ParseErrorBundle Text Void)
Arguments are: original input and parse error

newtype ParseExceptionBs

Parse exception thrown by mkURIBs when a given ByteString value cannot be parsed as a URI.

Since: 0.3.3.0

Constructors

ParseExceptionBs (ParseErrorBundle ByteString Void)
Arguments are: original input and parse error

Refined text

Refined text values can only be created by using the smart constructors listed below, such as mkScheme. This eliminates the possibility of having an invalid component in URI which could invalidate the whole URI.

Note that the refined text RText type is labelled at the type level with RTextLabels, which see.

When an invalid Text value is passed to a smart constructor, it rejects it by throwing the RTextException. Remember that the Maybe datatype is also an instance of MonadThrow, and so one could as well use the smart constructors in the Maybe monad.

data RText (l :: RTextLabel)

Refined text labelled at the type level.

data RTextLabel

Refined text labels.

Constructors

Scheme
See mkScheme

Host
See mkHost

Username
See mkUsername

Password
See mkPassword

PathPiece
See mkPathPiece

QueryKey
See mkQueryKey

QueryValue
See mkQueryValue

Fragment
See mkFragment

mkScheme :: MonadThrow m => Text -> m (RText 'Scheme)

Lift a Text value into RText Scheme.

Scheme names consist of a sequence of characters beginning with a letter and followed by any combination of letters, digits, plus "+", period ".", or hyphen "-".

This smart constructor performs normalization of valid schemes by converting them to lower case.

See also: https://tools.ietf.org/html/rfc3986#section-3.1

mkHost :: MonadThrow m => Text -> m (RText 'Host)

Lift a Text value into RText Host.

The host sub-component of authority is identified by an IP literal encapsulated within square brackets, an IPv4 address in dotted-decimal form, or a registered name.

This smart constructor performs normalization of valid hosts by converting them to lower case.

See also: https://tools.ietf.org/html/rfc3986#section-3.2.2

mkUsername :: MonadThrow m => Text -> m (RText 'Username)

Lift a Text value into RText Username.

This smart constructor does not perform any sort of normalization.

See also: https://tools.ietf.org/html/rfc3986#section-3.2.1

mkPassword :: MonadThrow m => Text -> m (RText 'Password)

Lift a Text value into RText Password.

This smart constructor does not perform any sort of normalization.

See also: https://tools.ietf.org/html/rfc3986#section-3.2.1

mkPathPiece :: MonadThrow m => Text -> m (RText 'PathPiece)

Lift a Text value into RText PathPiece.

This smart constructor does not perform any sort of normalization.

See also: https://tools.ietf.org/html/rfc3986#section-3.3

mkQueryKey :: MonadThrow m => Text -> m (RText 'QueryKey)

Lift a Text value into 'RText QueryKey.

This smart constructor does not perform any sort of normalization.

See also: https://tools.ietf.org/html/rfc3986#section-3.4

mkQueryValue :: MonadThrow m => Text -> m (RText 'QueryValue)

Lift a Text value into RText QueryValue.

This smart constructor does not perform any sort of normalization.

See also: https://tools.ietf.org/html/rfc3986#section-3.4

mkFragment :: MonadThrow m => Text -> m (RText 'Fragment)

Lift a Text value into RText Fragment.

This smart constructor does not perform any sort of normalization.

See also: https://tools.ietf.org/html/rfc3986#section-3.5

unRText :: RText l -> Text

Project a plain strict Text value from a refined RText l value.

data RTextException

The exception is thrown when a refined RText l value cannot be constructed due to the fact that given Text value is not correct.

Constructors

RTextException RTextLabel Text
RTextLabel identifying what sort of refined text value could not be constructed and the input that was supplied, as a Text value

Parsing

The input you feed into the parsers must be a valid URI as per RFC 3986, that is, its components should be percent-encoded where necessary.

parser :: MonadParsec e Text m => m URI

This parser can be used to parse URI from strict Text. Remember to use a concrete non-polymorphic parser type for efficiency.

parserBs :: MonadParsec e ByteString m => m URI

This parser can be used to parse URI from strict ByteString. Remember to use a concrete non-polymorphic parser type for efficiency.

Since: 0.0.2.0

Rendering

Rendering functions take care of constructing correct URI representation as per RFC 3986, that is, percent-encoding will be applied when necessary automatically.

render :: URI -> Text

Render a given URI value as strict Text.

render' :: URI -> Builder

Render a given URI value as a Builder.

renderBs :: URI -> ByteString

Render a given URI value as a strict ByteString.

renderBs' :: URI -> Builder

Render a given URI value as a Builder.

renderStr :: URI -> String

Render a given URI value as a String.

Since: 0.0.2.0

renderStr' :: URI -> ShowS

Render a given URI value as ShowS.

Since: 0.0.2.0