Text.URI.QQ

Description

Quasi-quoters for compile-time construction of URIs and refined text values.

All of the quasi-quoters in this module can be used in an expression context. With the ViewPatterns language extension enabled, they may also be used in a pattern context (since 0.3.2.0).

Documentation

uri :: QuasiQuoter

Construct a URI value at compile time.

scheme :: QuasiQuoter

Construct a RText Scheme value at compile time.

host :: QuasiQuoter

Construct a RText Host value at compile time.

username :: QuasiQuoter

Construct a RText Username value at compile time.

password :: QuasiQuoter

Construct a RText Password value at compile time.

pathPiece :: QuasiQuoter

Construct a RText PathPiece value at compile time.

queryKey :: QuasiQuoter

Construct a RText QueryKey value at compile time.

queryValue :: QuasiQuoter

Construct a 'RText QueryValue value at compile time.

fragment :: QuasiQuoter

Construct a RText Fragment value at compile time.